+++ packages/react-resizable-panels-website/tests/Storage.spec.ts
@@ -3,32 +3,26 @@ import { createElement } from "react";
 import { Panel, PanelGroup, PanelResizeHandle } from "react-resizable-panels";
 
 import { verifyAriaValues } from "./utils/aria";
+import { imperativeCollapsePanel, imperativeExpandPanel } from "./utils/panels";
 import { goToUrl } from "./utils/url";
 
 const panelGroupABC = createElement(
   PanelGroup,
   { autoSaveId: "test-group", direction: "horizontal" },
-  createElement(Panel, { minSize: 10, order: 1 }),
+  createElement(Panel, {
+    minSize: 10,
+    order: 1,
+  }),
   createElement(PanelResizeHandle),
-  createElement(Panel, { minSize: 10, order: 2 }),
+  createElement(Panel, {
+    minSize: 10,
+    order: 2,
+  }),
   createElement(PanelResizeHandle),
-  createElement(Panel, { minSize: 10, order: 3 })
-);
-
-const panelGroupBC = createElement(
-  PanelGroup,
-  { autoSaveId: "test-group", direction: "horizontal" },
-  createElement(Panel, { minSize: 10, order: 2 }),
-  createElement(PanelResizeHandle),
-  createElement(Panel, { minSize: 10, order: 3 })
-);
-
-const panelGroupAB = createElement(
-  PanelGroup,
-  { autoSaveId: "test-group", direction: "horizontal" },
-  createElement(Panel, { minSize: 10, order: 1 }),
-  createElement(PanelResizeHandle),
-  createElement(Panel, { minSize: 10, order: 2 })
+  createElement(Panel, {
+    minSize: 10,
+    order: 3,
+  })
 );
 
 test.describe("Storage", () => {
@@ -64,8 +58,7 @@ test.describe("Storage", () => {
         now: 80,
       });
 
-      // Wait for localStorage write debounce
-      await new Promise((resolve) => setTimeout(resolve, 250));
+      await waitForLocalStorageWrite();
 
       // Values should be remembered after a page reload
       await page.reload();
@@ -80,6 +73,22 @@ test.describe("Storage", () => {
     test("should store layouts separately per panel combination", async ({
       page,
     }) => {
+      const panelGroupBC = createElement(
+        PanelGroup,
+        { autoSaveId: "test-group", direction: "horizontal" },
+        createElement(Panel, { minSize: 10, order: 2 }),
+        createElement(PanelResizeHandle),
+        createElement(Panel, { minSize: 10, order: 3 })
+      );
+
+      const panelGroupAB = createElement(
+        PanelGroup,
+        { autoSaveId: "test-group", direction: "horizontal" },
+        createElement(Panel, { minSize: 10, order: 1 }),
+        createElement(PanelResizeHandle),
+        createElement(Panel, { minSize: 10, order: 2 })
+      );
+
       await goToUrl(page, panelGroupABC);
 
       const resizeHandles = page.locator("[data-panel-resize-handle-id]");
@@ -93,8 +102,7 @@ test.describe("Storage", () => {
         now: 33,
       });
 
-      // Wait for localStorage write debounce
-      await new Promise((resolve) => setTimeout(resolve, 250));
+      await waitForLocalStorageWrite();
 
       // Hide the first panel and then resize things
       await goToUrl(page, panelGroupBC);
@@ -104,8 +112,7 @@ test.describe("Storage", () => {
         now: 10,
       });
 
-      // Wait for localStorage write debounce
-      await new Promise((resolve) => setTimeout(resolve, 250));
+      await waitForLocalStorageWrite();
 
       // Hide the last panel and then resize things
       await goToUrl(page, panelGroupAB);
@@ -115,8 +122,7 @@ test.describe("Storage", () => {
         now: 90,
       });
 
-      // Wait for localStorage write debounce
-      await new Promise((resolve) => setTimeout(resolve, 250));
+      await waitForLocalStorageWrite();
 
       // Reload and verify all of the different layouts are remembered individually
       await goToUrl(page, panelGroupABC);
@@ -137,5 +143,105 @@ test.describe("Storage", () => {
         now: 90,
       });
     });
+
+    test("should remember the most recent expanded size for collapsed panels", async ({
+      page,
+    }) => {
+      const panelGroup = createElement(
+        PanelGroup,
+        { autoSaveId: "test-group", direction: "horizontal" },
+        createElement(Panel, {
+          collapsible: true,
+          id: "left",
+          minSize: 10,
+          order: 1,
+        }),
+        createElement(PanelResizeHandle),
+        createElement(Panel, {
+          collapsible: true,
+          id: "middle",
+          minSize: 10,
+          order: 2,
+        }),
+        createElement(PanelResizeHandle),
+        createElement(Panel, {
+          collapsible: true,
+          id: "right",
+          minSize: 10,
+          order: 3,
+        })
+      );
+
+      await goToUrl(page, panelGroup);
+
+      const resizeHandles = page.locator("[data-panel-resize-handle-id]");
+      const first = resizeHandles.first();
+      const last = resizeHandles.last();
+
+      await verifyAriaValues(first, {
+        now: 33,
+      });
+      await verifyAriaValues(last, {
+        now: 33,
+      });
+
+      // Change panel sizes
+      await first.focus();
+      await page.keyboard.press("ArrowLeft");
+      await last.focus();
+      await page.keyboard.press("ArrowRight");
+
+      // Verify sizes
+      await verifyAriaValues(first, {
+        now: 23,
+      });
+      await verifyAriaValues(last, {
+        now: 53,
+      });
+
+      await waitForLocalStorageWrite();
+
+      // Collapse panels
+      await imperativeCollapsePanel(page, "left");
+      await imperativeCollapsePanel(page, "right");
+
+      // Verify sizes
+      await verifyAriaValues(first, {
+        now: 0,
+      });
+      await verifyAriaValues(last, {
+        now: 100,
+      });
+
+      await waitForLocalStorageWrite();
+
+      // Reload page
+      await page.reload();
+
+      // Verify collapsed sizes resized
+      await verifyAriaValues(first, {
+        now: 0,
+      });
+      await verifyAriaValues(last, {
+        now: 100,
+      });
+
+      // Expand panels
+      await imperativeExpandPanel(page, "left");
+      await imperativeExpandPanel(page, "right");
+
+      // Verify sizes
+      await verifyAriaValues(first, {
+        now: 23,
+      });
+      await verifyAriaValues(last, {
+        now: 53,
+      });
+    });
   });
 });
+
+// Wait for localStorage write debounce
+async function waitForLocalStorageWrite() {
+  await new Promise((resolve) => setTimeout(resolve, 250));
+}